feat: logger winston and Errors

+++ src/resources/tasks/task.router.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
 const tasksService = require('./task.service');
 const catchErrors = require('../../common/catchErrors');
+const createError = require('http-errors');
+const validator = require('validator');
 
 router.route('/').get(
   catchErrors(async (req, res) => {
@@ -11,9 +13,12 @@ router.route('/').get(
 
 router.route('/:id').get(
   catchErrors(async (req, res) => {
+    if (!validator.isUUID(req.params.id)) {
+      throw createError(400, `Task ID'${req.params.id}' not UUID`);
+    }
     const task = await tasksService.getTaskId(req.boardId, req.params.id);
     if (!task) {
-      return res.status(404).json({ message: 'Not found' });
+      throw createError(404, `Task '${req.params.id}' not found`);
     }
     return res.status(200).json(task);
   })
@@ -21,11 +26,17 @@ router.route('/:id').get(
 
 router.route('/:id').put(
   catchErrors(async (req, res) => {
+    if (!validator.isUUID(req.params.id)) {
+      throw createError(400, `Task ID'${req.params.id}' not UUID`);
+    }
     const task = await tasksService.editTask(
       req.boardId,
       req.params.id,
       req.body
     );
+    if (!task) {
+      throw createError(404, `Task '${req.params.id}' not found`);
+    }
     return res.status(200).json(task);
   })
 );
@@ -39,8 +50,14 @@ router.route('/').post(
 
 router.route('/:id').delete(
   catchErrors(async (req, res) => {
+    if (!validator.isUUID(req.params.id)) {
+      throw createError(400, `Task ID'${req.params.id}' not UUID`);
+    }
     const message = await tasksService.deleteTask(req.params.id);
-    return res.status(200, 204).json({ message });
+    if (!message) {
+      throw createError(404, `Task '${req.params.id}' not found`);
+    }
+    return res.status(204).json({ message });
   })
 );
